[android-11.0.0_r20] Fix file size inconsistency across upper and lower fs
When mixing IO via ContentResolver and file paths, there's a risk of
the upper fs (via file path) not seeing the size change from a write
directly to the lower fs (via ContentResolver).

This is a limitation from the kernel handling attrs with write back
cache enabled and writes to the lower fs directly.

We now workaround and force eventual consistency after the lower fs fd
is closed.

Test: Manual
Bug: 161228757
Bug: 161767712
Change-Id: I6dfaaf89df94235b03ed4c9c85ecc8c4b39b44ff